break out funcs to libraries
