+++ tree.py
@@ -1,6 +1,7 @@
 """
 inspiration from https://github.com/lianemeth/forest/blob/master/forest/NaryTree.py
 """
+import config
 import locale
 import weakref
